move helper tests to helpers_test.go
